Modified MH's branch to add global song info storage.
Modified MH's branch to add global song info storage. This means that designers and other programmers can simply specify a song by a user-friendly name elsewhere in the code and the MusicDictionary can take care of automatically telling the rhythm music player exactly what song it is and how fast it plays.