Revamp of UI to be closer to new requirement.
Revamped the UI to be closer in capability to Hi Fi Rush as per given instructions. A prefab is included that will be worked on for the final project.
FIXME: There is a nasty non-deterministic bug occurring because of Unity start up lag. This means the music and UI may not sync up properly. Ongoing support for fix soon!