Boss AI implementation testing
I got the work Nathan and Bryan did and put it into my RT Scene to test. I also outlined a BossAI script. My job right now is to give direction to the mechanics that a team of programmers are creating. I've given my feedback and next week we will continue to work on the mechanics since they're not quite ready.