Enemy Hitbox Update
Made a solid base for enemy hitboxes that include shapes and derives from data assets, like the player one. Have tested it only for a bit and it seems to be functioning as far as visualizing it but no gameplay functions yet. Need to test it with HealthComponent, which is what is in the code already just need to it to work.