Boss Prefab First Pass Done
I'm gonna merge request this pass. The boss is mostly working. I need to wait on some work from the boss team and I need to implement a proper player controller and enemy prefabs for the summoning in my scene. Aside from that I am pretty content with the progress today and I think I can finish this up in class on monday with some help and feedback.