Boss Health and MGR Mechanic
I migrated the boss health script and its logic to our global health script so the boss can be damaged. Also created the MGR mechanic which has a uses MGRProjectile prefab and 2 new scripts called MachineGunRiff and MGRProjectile. I also remade the boss prefab for dropping into scenes if need be but some of the other mechanics need testing and you need to assign the player game object in the scripts that need it.