Blackout Fake Stairways & Misc Aesthetic fixes
- Dark Stairways:

Created darkened fake stairways that improve the look of stairs leading to nowhere. Used them for all of the teleporting areas.

Misc Aesthetics: Fixed some of the gaps in the level between planes and walls, and added some additional geometry to improve overall look of the level.

Find in the scene folder under Nicolas Salman.